convert submissions to full server rendering
