+++ app/submissions/Submissions.js
@@ -3,20 +3,30 @@ import { headers } from 'next/headers'
 import api from '../../lib/api-client'
 import { prettyId } from '../../lib/utils'
 import serverAuth from '../auth'
-import SubmissionList from './SubmissionList'
 import LoadingSpinner from '../../components/LoadingSpinner'
+import NoteList from '../../components/NoteList'
+import PaginationLinks from '../../components/PaginationLinks'
 
-export default async function Submissions({ groupId, invitationId }) {
+const displayOptions = {
+  pdfLink: false,
+  htmlLink: false,
+  showContents: false,
+  emptyMessage: 'No submissions found.',
+}
+
+export default async function Submissions({ groupId, invitationId, page }) {
   const { token } = await serverAuth()
   const headersList = await headers()
   const remoteIpAddress = headersList.get('x-forwarded-for')
   const notesPerPage = 25
+  const currentPage = Math.max(parseInt(page, 10) || 1, 1)
 
-  const getFirstPageNotesP = api.get(
+  const { notes, count } = await api.get(
     '/notes',
     {
       invitation: invitationId,
       limit: notesPerPage,
+      offset: notesPerPage * (currentPage - 1),
     },
     {
       accessToken: token,
@@ -31,10 +41,15 @@ export default async function Submissions({ groupId, invitationId }) {
         <hr />
       </header>
       <Suspense fallback={<LoadingSpinner />}>
-        <SubmissionList
-          getFirstPageNotesP={getFirstPageNotesP}
-          notesPerPage={notesPerPage}
-          invitationId={invitationId}
+        <div className="notes">
+          <NoteList notes={notes} displayOptions={displayOptions} />
+        </div>
+        <PaginationLinks
+          currentPage={currentPage}
+          itemsPerPage={notesPerPage}
+          totalCount={count}
+          baseUrl="/submissions"
+          queryParams={{ venue: groupId }}
         />
       </Suspense>
     </>
