+++ components/Collapse.js
@@ -1,3 +1,5 @@
+'use client'
+
 /* globals $: false */
 import { nanoid } from 'nanoid'
 import { useRef, useState } from 'react'
