+++ components/NoteContent.js
@@ -1,3 +1,5 @@
+'use client'
+
 /* globals DOMPurify,marked: false */
 
 import { useState, useEffect } from 'react'
