+++ components/UnlinkPublicationButton.js
@@ -1,3 +1,5 @@
+'use client'
+
 import { useState } from 'react'
 import Icon from './Icon'
